requirements.txt
fastapi==0.109.2
uvicorn==0.27.0.post1
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
httpx==0.27.0
pydantic==2.6.1
python-dotenv==1.0.1
